Add asserts around Swift singletons.
